refactor

+++ engine/components/directional-light.hpp
@@ -1,9 +1,10 @@
 #pragma once
-#include "math.hpp"
-#include "math/intersection.hpp"
-#include "math/ray.hpp"
+#include "core/math.hpp"
+#include "core/math/intersection.hpp"
+#include "core/math/ray.hpp"
 #include "render/light-data.hpp"
 #include "render/material.hpp"
+#include "render/light-common.hpp"
 #include <functional>
 #include <math.h>
 namespace engine::components
@@ -20,7 +21,7 @@ namespace engine::components
             {
                 return;
             }
-            float specular = pow(dot(light_data.normal, light_data.view_dir), mat.glossiness) * mat.specular;
+            float specular = std::powf(dot(light_data.normal, light_data.view_dir), mat.glossiness) * mat.specular;
             light_data.color += color * ndotl * (ndotl * mat.albedo + specular);
         }
     };
